enable auth for notes

+++ BE/src/controllers/notes.ts
@@ -2,6 +2,7 @@ import { NextFunction, RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../utils/assertIsDefined";
 
 interface CreateNoteBody {
   title?: string;
@@ -22,8 +23,10 @@ export const getNotes: RequestHandler = async (
   res,
   next: NextFunction
 ) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -32,13 +35,18 @@ export const getNotes: RequestHandler = async (
 
 export const getNote: RequestHandler = async (req, res, next: NextFunction) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(400, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
     return res.status(200).json(note);
   } catch (error) {
@@ -52,13 +60,19 @@ export const createNote: RequestHandler<
   CreateNoteBody,
   unknown
 > = async (req, res, next: NextFunction) => {
+  const authenticatedUserId = req.session.userId;
   const title = req.body.title;
   const text = req.body.text;
   try {
+    assertIsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
-    const newNote = await NoteModel.create({ title, text });
+    const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
+      title,
+      text,
+    });
     res.status(201).json(newNote);
   } catch (error) {
     next(error);
@@ -71,10 +85,12 @@ export const updateNote: RequestHandler<
   UpdateNoteBody,
   unknown
 > = async (req, res, next: NextFunction) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
@@ -85,6 +101,9 @@ export const updateNote: RequestHandler<
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(400, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     note.title = newTitle;
@@ -103,14 +122,19 @@ export const deleteNote: RequestHandler<
   unknown,
   unknown
 > = async (req, res, next: NextFunction) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     await note.deleteOne();
